Move Phpunit test calls to a separate class.
